refactor: Payment 도메인 분리로 Point 테스트코드 수정, TestBase에 PaymentController 추가
