+++ src/test/java/com/example/surveyapp/config/testbase/WebMvcTestBase.java
@@ -3,6 +3,7 @@ package com.example.surveyapp.config.testbase;
 import com.example.surveyapp.domain.admin.presentation.AdminController;
 import com.example.surveyapp.domain.ai.chat.presentation.ChatController;
 import com.example.surveyapp.domain.order.presentation.OrderController;
+import com.example.surveyapp.domain.payment.presentation.PaymentController;
 import com.example.surveyapp.domain.point.presentation.PointController;
 import com.example.surveyapp.domain.product.presentation.ProductController;
 import com.example.surveyapp.domain.survey.presentation.OptionsController;
@@ -26,7 +27,7 @@ import org.springframework.test.web.servlet.MockMvc;
         controllers = {UserController.class, AdminController.class, OrderController.class,
                 PointController.class, ProductController.class, OptionsController.class,
                 QuestionController.class, SurveyController.class, SurveyAnswerController.class,
-                ChatController.class},
+                ChatController.class, PaymentController.class},
         excludeFilters = @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = JwtFilter.class)
 )
 @AutoConfigureMockMvc(addFilters = false)
